Move the depth conversion code into x11rb
Nothing was actually changed yet.
