+++ examples/display_ppm.rs
@@ -1,12 +1,9 @@
-use std::borrow::Cow;
-use std::convert::TryInto;
-
 use x11rb::connection::Connection;
-use x11rb::errors::{ParseError, ReplyOrIdError};
-use x11rb::image::Image;
+use x11rb::errors::ReplyOrIdError;
+use x11rb::image::{ColorComponent, Image, PixelLayout, convert_image};
 use x11rb::protocol::xproto::{
-    AtomEnum, ConnectionExt, CreateGCAux, CreateWindowAux, PropMode, Screen, Setup, VisualClass,
-    Visualid, Visualtype, Window, WindowClass,
+    AtomEnum, ConnectionExt, CreateGCAux, CreateWindowAux, PropMode, Screen, VisualClass, Visualid,
+    Window, WindowClass,
 };
 use x11rb::protocol::Event;
 use x11rb::wrapper::ConnectionExt as _;
@@ -15,212 +12,6 @@ x11rb::atom_manager! {
     Atoms: AtomsCookie {
         WM_PROTOCOLS,
         WM_DELETE_WINDOW,
-    }
-}
-
-/// The description of a single color component.
-///
-/// For example, in an RGB image, pixels are often saved as `0xRRGGBB`, where each letter
-/// represents the respective color component. In the example, green has a `width` of 8 (it takes
-/// up 8 bits) and a `shift` of `16` (there are 16 less significant bits beyond it). This info is
-/// represented as a `ColorComponent`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct ColorComponent {
-    /// Number of bits for the component
-    width: u8,
-    /// Offset in an u32 of the component
-    shift: u8,
-}
-
-impl ColorComponent {
-    /// Create a new color component with the given information.
-    pub fn new(width: u8, shift: u8) -> Self {
-        assert!(
-            width < 16,
-            "At most 16 bits for a single color component supported, got {}",
-            width,
-        );
-        assert!(
-            shift < 32,
-            "Too large shift for color component, got {}, but pixels are 32 bit",
-            shift,
-        );
-        assert!(
-            shift + width <= 32,
-            "Would overflow a 32 bit value when putting {} bits at offset {}",
-            width,
-            shift,
-        );
-        Self { width, shift }
-    }
-
-    /// Get the pixel mask representing this color component.
-    ///
-    /// The mask can be used to mask off other colors in a pixel value. Only the bits that
-    /// correspond to this color component are set.
-    /// ```
-    /// let red = ColorComponent::new(8, 16);
-    /// assert_eq(red.to_mask(), 0xff0000);
-    /// ```
-    fn to_mask(self) -> u32 {
-        // Get a mask with 'width' set bits.
-        let mask = (1u32 << self.width) - 1;
-        // Shift the mask into the right position
-        mask << self.shift
-    }
-
-    /// Create a new color component from a color mask.
-    ///
-    /// This turns a color mask into its individual components.
-    /// ```
-    /// let red1 = ColorComponent::new(8, 16);
-    /// let red2 = ColorComponent::from_mask(0xff0000);
-    /// ```
-    pub fn from_mask(mask: u32) -> Self {
-        let width = mask.count_ones();
-        let shift = mask.trailing_zeros();
-        // Both width and shift can be at most 32, which should fit into u8.
-        let result = Self::new(width.try_into().unwrap(), shift.try_into().unwrap());
-        assert_eq!(mask, result.to_mask(), "Malformed mask {}", mask);
-        result
-    }
-
-    /// Get this color component out of a pixel value.
-    ///
-    /// This function gets a single pixel value and returns this color component's value in that
-    /// pixel value, expanded to width 16.
-    /// ```
-    /// let red = ColorComponent::new(8, 16);
-    /// assert_eq!(0xABAB, red.decode(0x78AB_4321));
-    /// panic!("are these even run?");
-    /// ```
-    pub fn decode(self, pixel: u32) -> u16 {
-        // Get the color component out
-        let value = (pixel & self.to_mask()) >> self.shift;
-
-        // Now expand from with `self.width` to width 16.
-        let mut width = self.width;
-        let mut value = value << (16 - width);
-        // Add some low bits by using the high bits
-        while width < 16 {
-            value |= value >> width;
-            width <<= 1;
-        }
-        value.try_into().unwrap()
-    }
-
-    /// Encode a color value according to this pixel format.
-    ///
-    /// ```
-    /// let red = ColorComponent::new(8, 16);
-    /// assert_eq!(0xAB0000, red.encode(0xABCD));
-    /// ```
-    pub fn encode(self, intensity: u16) -> u32 {
-        // First truncate to width `self.width`, then place at the correct offset.
-        (u32::from(intensity) >> (16 - self.width)) << self.shift
-    }
-}
-
-/// A collection of color components describing the red, green, and blue components of a pixel.
-///
-/// A [`ColorComponent`] describes a single color component in an image. This structure describes
-/// the `red`, `green`, and `blue` color components by containing a [`ColorComponent`] for each of
-/// them.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct PixelLayout {
-    red: ColorComponent,
-    green: ColorComponent,
-    blue: ColorComponent,
-}
-
-impl PixelLayout {
-    /// Create a new pixel layout from the description of each component
-    pub fn new(red: ColorComponent, green: ColorComponent, blue: ColorComponent) -> Self {
-        Self { red, green, blue }
-    }
-
-    /// Create a new pixel layout
-    pub fn from_visual_type(visual: Visualtype) -> Self {
-        if visual.class != VisualClass::TrueColor && visual.class != VisualClass::DirectColor {
-            panic!("Visual does not use decomposed colors: {:x?}", visual);
-        }
-        Self::new(
-            ColorComponent::from_mask(visual.red_mask),
-            ColorComponent::from_mask(visual.green_mask),
-            ColorComponent::from_mask(visual.blue_mask),
-        )
-    }
-
-    /// Get the depth of this pixel layout.
-    ///
-    /// The depth is the number of significant bits of each pixel value.
-    pub fn depth(self) -> u8 {
-        // TODO: I am not quite sure this implementation is correct. The protocol seems to allow
-        // unused bits in the middle..?
-        self.red.width + self.green.width + self.blue.width
-    }
-
-    /// Decode a pixel value into its red, green, and blue components.
-    ///
-    /// This function returns each component expanded to width 16.
-    /// ```
-    /// let layout = PixelLayout::new(
-    ///     ColorComponent::new(8, 16),
-    ///     ColorComponent::new(8, 8),
-    ///     ColorComponent::new(8, 0),
-    /// );
-    /// assert_eq!((0xABAB, 0x4343, 0x2121), layout.decode(0x78AB_4321));
-    /// ```
-    pub fn decode(self, pixel: u32) -> (u16, u16, u16) {
-        let red = self.red.decode(pixel);
-        let green = self.green.decode(pixel);
-        let blue = self.blue.decode(pixel);
-        (red, green, blue)
-    }
-
-    /// Encode a color value according to this layout.
-    ///
-    /// ```
-    /// let layout = PixelLayout::new(
-    ///     ColorComponent::new(8, 16),
-    ///     ColorComponent::new(8, 8),
-    ///     ColorComponent::new(8, 0),
-    /// );
-    /// assert_eq!(0x78AB_4321, layout.encode((0xABAB, 0x4343, 0x2121)));
-    /// ```
-    pub fn encode(self, (red, green, blue): (u16, u16, u16)) -> u32 {
-        let red = self.red.encode(red);
-        let green = self.green.encode(green);
-        let blue = self.blue.encode(blue);
-        red | green | blue
-    }
-}
-
-/// Convert the given `image` to another layout and storage.
-///
-/// This function gets an `image` with pixel values that can be interpreted according to
-/// `input_layout`. It produces an image with pixel values according to `output_layout`. The
-/// resulting image is stored in the native format of the X11 server described by `setup`.
-fn convert_image<'image, 'data>(
-    image: &'image Image<'data>,
-    setup: &Setup,
-    input_layout: PixelLayout,
-    output_layout: PixelLayout,
-) -> Result<Cow<'image, Image<'data>>, ParseError> {
-    if input_layout == output_layout {
-        image.native(setup)
-    } else {
-        // Yay, we get to convert the image :-(
-        let (width, height) = (image.width(), image.height());
-        let mut result = Image::allocate_native(width, height, output_layout.depth(), setup)?;
-        for y in 0..height {
-            for x in 0..width {
-                let pixel = image.get_pixel(x, y);
-                let pixel = output_layout.encode(input_layout.decode(pixel));
-                result.put_pixel(x, y, pixel);
-            }
-        }
-        Ok(Cow::Owned(result))
     }
 }
